Test fixes.

+++ app/Http/Controllers/VenuesController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use App\Http\Requests\VenueCreateFormRequest;
 use App\Http\Requests\VenueEditFormRequest;
 use App\Models\Venue;
+use Illuminate\Http\Request;
+
 
 class VenuesController extends Controller
 {
@@ -109,6 +111,8 @@ class VenuesController extends Controller
      */
     public function destroy(Venue $venue)
     {
+        $this->authorize('delete', Venue::class);
+
         $venue->delete();
 
         return redirect()->route('venues.index');
